refactor typechecking into modules
